added conversions for postgres binary mode

+++ src/conversion.utils.hpp
@@ -21,6 +21,6 @@
 #include <string>
 
 std::string itos(const int i);
-double ntohll(const uint64_t x);
+uint64_t ntohll(const uint64_t x);
 
 #endif // CONVERSION_UTILS_HPP
